backend/requirements.txt
# FastAPI & Server
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# Database (using psycopg with binary wheels - no compilation needed)
sqlalchemy[asyncio]>=2.0.0
psycopg[binary]>=3.1.0
alembic>=1.12.0

# HTTP & Scraping
httpx>=0.25.0
beautifulsoup4>=4.12.0

# Validation & Settings
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
